+++ src/Symfony/Component/Workflow/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 4.0.0
 -----
 
+ * Added TransitionBlockers as a way to convey why exactly transitions can't be made (#24501).
  * Removed class name support in `WorkflowRegistry::add()` as second parameter.
 
 3.4.0